+++ src/constants/okx.ts
@@ -5,3 +5,7 @@ export const LangMap = {
     [Locale.zhHans]: 'zh_cn',
     [Locale.zhHant]: 'zh_tw',
 };
+
+export const OKX_HOST = 'https://okx-proxy.r2d2.to';
+
+export const NATIVE_TOKEN_ADDRESS = '0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE';
